+++ comet/interactive.py
@@ -1,7 +1,7 @@
 import logging
 import argparse
 
-from src.comet_model import PretrainedCometModel
+from comet.comet_model import PretrainedCometModel
 
 logging.basicConfig(
     format='%(asctime)s - %(levelname)s - %(name)s -   %(message)s', datefmt='%m/%d/%Y %H:%M:%S', level=logging.INFO)
